Extract proxy constants and bridge DNS management into ProxyConfig
Move port constants, domain sets, gateway IP resolution, DNS override
management, and Vertex host resolution from MitmProxy into a new
ProxyConfig class. MitmProxy retains thin delegating methods for
backward compatibility.

This separates Vert.x-independent proxy configuration (needed by CLI
commands like init, build, doctor) from the Vert.x-dependent proxy
server, preparing for a future split into separate CLI and proxy
binaries.
